moved util/socket to ws_client
